+++ mxnet/src/main/java/org/bytedeco/mxnet/presets/mxnet.java
@@ -43,7 +43,7 @@ import org.bytedeco.opencv.presets.*;
 @Properties(inherit = {openblas.class, opencv_imgcodecs.class, opencv_highgui.class}, target = "org.bytedeco.mxnet", global = "org.bytedeco.mxnet.global.mxnet", value = {
     @Platform(value = {"linux", "macosx", "windows"}, compiler = {"cpp11", "fastfpu"},
         define = {"DMLC_USE_CXX11 1", "MSHADOW_USE_CBLAS 1", "MSHADOW_IN_CXX11 1", "MSHADOW_USE_CUDA 0", "MSHADOW_USE_F16C 0", "MXNET_USE_TVM_OP 0"},
-        include = {"mxnet/c_api.h", "mxnet/c_predict_api.h", /*"dmlc/base.h", "dmlc/io.h", "dmlc/logging.h", "dmlc/type_traits.h",
+        include = {"mxnet/c_api.h", "mxnet/c_predict_api.h", "nnvm/c_api.h", /*"dmlc/base.h", "dmlc/io.h", "dmlc/logging.h", "dmlc/type_traits.h",
                    "dmlc/parameter.h", "mshadow/base.h", "mshadow/expression.h", "mshadow/tensor.h", "mxnet/base.h",*/
                    "org_apache_mxnet_init_native_c_api.cc", "org_apache_mxnet_native_c_api.cc"},
         link = "mxnet", preload = {"mkldnn@.1", "libmxnet"}, /*resource = {"include", "lib"},*/
@@ -108,7 +108,7 @@ public class mxnet implements LoadEnabled, InfoMapper {
 
     public void map(InfoMap infoMap) {
         infoMap.put(new Info("org_apache_mxnet_init_native_c_api.cc", "org_apache_mxnet_native_c_api.cc").skip())
-               .put(new Info("MXNET_EXTERN_C", "MXNET_DLL").cppTypes().annotations())
+               .put(new Info("MXNET_EXTERN_C", "MXNET_DLL", "NNVM_DLL").cppTypes().annotations())
                .put(new Info("MSHADOW_USE_F16C", "MXNET_USE_TVM_OP").define(false))
                .put(new Info("MXNDArrayCreateSparseEx64", "MXNDArrayGetAuxNDArray64", "MXNDArrayGetAuxType64",
                              "MXNDArrayGetShape64", "MXSymbolInferShape64", "MXSymbolInferShapePartial64").skip())